requirements.txt
telethon>=1.28.0
langchain>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.21.0 
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain-community>=0.1.0
sentence-transformers>=2.5.1